updated DB - new table - facebook_friends; updated db logic WorkerThreads are now with limits for amount of threads

+++ DbSnappers/src/com/emerginggames/snappers/data/LevelTable.java
@@ -76,7 +76,7 @@ public class LevelTable  extends SQLiteTable<Level>{
         }
     }
     
-    public int countLevels(int packId){
+    public int countLevels(long packId){
         return count(String.format("%s = %d", KEY_LEVEL_PACK_ID, packId));
     }
